+++ apps/iOS/Sources/BrevApp.swift
@@ -74,6 +74,7 @@ struct BrevApp: App {
                         initialCurrentAccountID: session.backend?.account.id,
                         mailboxContext: settingsMailboxContext,
                         backendProvider: { accountID in session.backends[accountID] },
+                        pimSourceCoordinator: session.pimSourceCoordinator,
                         onAddAccount: { isShowingAddAccountSheet = true },
                         onSignOut: { account in await session.signOut(account: account) },
                         onRemoveAccount: { account in await session.removeAccount(account) },